Update profiles, payment calculations and add new tables
Updated the profiles.yml to use environment variables for passwords. Enhanced payment calculations to include USD equivalents by joining with the prices table. Added a new incremental table for mints and simplified the mint per month query. Added a basic index.html file with an iframe for embedding a dashboard.